Update 31 August 2023 - Module Legality Type (Done) - Update type data field active to tinyInteger

+++ database/migrations/2023_07_23_061222_create_units_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('munits', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name',150);
-            $table->integer('active')->default('1');
+            $table->tinyInteger('active')->default('1');
             $table->timestamps();
         });
     }
